Refactoring lambda manager and plotting tool.

+++ lambda-manager/src/main/java/com/lambda_manager/utils/CleanLogFiles.java
@@ -5,7 +5,9 @@ import io.micronaut.runtime.event.ApplicationStartupEvent;
 
 import javax.inject.Singleton;
 import java.io.File;
-import java.util.Objects;
+import java.io.FilenameFilter;
+
+import static com.lambda_manager.utils.Constants.*;
 
 @SuppressWarnings("unused")
 @Singleton
@@ -13,13 +15,23 @@ public class CleanLogFiles implements ApplicationEventListener<ApplicationStartu
 
     @Override
     public void onApplicationEvent(ApplicationStartupEvent event) {
-        File lambdasDir = new File("src/lambdas");
-        for (File lambdaDir : Objects.requireNonNull(lambdasDir.listFiles())) {
-            for (File logDir : Objects.requireNonNull(lambdaDir.listFiles(pathname -> pathname.getName().contains("logs")))) {
-                for (File file : Objects.requireNonNull(logDir.listFiles())) {
-                    //noinspection ResultOfMethodCallIgnored
-                    file.delete();
+        purgeDirectory(new File(CODEBASE), false);
+        purgeDirectory(new File(LAMBDA_LOGS), false);
+        purgeDirectory(new File(MANAGER_LOGS), false);
+    }
+
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    private void purgeDirectory(File dir, boolean deleteDir) {
+        File[] files = dir.listFiles((dir1, name) -> !name.equals(".gitkeep"));
+        if(files != null) {
+            for (File file: files) {
+                if (file.isDirectory()) {
+                    purgeDirectory(file, true);
                 }
+                file.delete();
+            }
+            if(deleteDir) {
+                dir.delete();
             }
         }
     }
